% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Add content to slide candidate}
\usage{
add(
  robopptx,
  content = NULL,
  robo_class = NA_character_,
  r_class = NA_character_
)
}
\arguments{
\item{robopptx}{robopptx Imported layout file from \link{load_layout}}

\item{content}{object Content to add to the current slide candidate}
}
\value{
A \code{robopptx} object.
}
\description{
Add content to slide candidate
}
\examples{
"todo"
}
