% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robopptx.R
\name{add_alignment_layout_table}
\alias{add_alignment_layout_table}
\title{Add alignment to layout table}
\usage{
add_alignment_layout_table(layout_table, position_precision = 0.5)
}
\arguments{
\item{layout_table}{data.table describing layouts returned by \link{get_layout_table}}
}
\value{
data.table
}
\description{
Add alignment to layout table
}
\examples{
"todo"
}
\keyword{internal}
