% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_footer}
\alias{add_footer}
\title{Add Text to slide candidate}
\usage{
add_footer(robopptx, footer = NULL)
}
\arguments{
\item{robopptx}{robopptx Imported layout file from \link{load_layout}}

\item{footer}{character Footer to add to the current slide candidate}
}
\value{
A \code{robopptx} object.
}
\description{
Add Text to slide candidate
}
\examples{
"todo"
}
