% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_graph}
\alias{add_graph}
\title{Add Graph to slide candidate}
\usage{
add_graph(robopptx, graph = NULL)
}
\arguments{
\item{robopptx}{robopptx Imported layout file from \link{load_layout}}

\item{graph}{ggplot, mschart, charater Graph to add to the current slide candidate as ggplot, mschart or filepath to existing .img or .png file}
}
\value{
A \code{robopptx} object.
}
\description{
Add Graph to slide candidate
}
\examples{
"todo"
}
