% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robopptx.R
\name{add_layoutcombination_layout_table}
\alias{add_layoutcombination_layout_table}
\title{Add Layout Combination of Contents to layout Table}
\usage{
add_layoutcombination_layout_table(layout_table)
}
\arguments{
\item{layout_table}{data.table describing layouts returned by \link{get_layout_table}}
}
\value{
character name of grouped class
}
\description{
Add Layout Combination of Contents to layout Table
}
\examples{
"todo"
}
\keyword{internal}
