% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robopptx.R
\name{add_placementorder_layout_table}
\alias{add_placementorder_layout_table}
\title{Add Order of every placement per Layout}
\usage{
add_placementorder_layout_table(layout_table)
}
\arguments{
\item{layout_table}{data.table describing layouts returned by \link{get_layout_table}}
}
\value{
character name of grouped class
}
\description{
Add Order of every placement per Layout
}
\examples{
add_placementorder_layout_table(get_layout_table())
}
\keyword{internal}
