% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robopptx.R
\name{add_robocop_variables}
\alias{add_robocop_variables}
\title{Add necessary variables for robocop to layer table}
\usage{
add_robocop_variables(layout_table, position_precision = 0.5)
}
\arguments{
\item{layout_table}{data.table describing layouts returned by \link{get_layout_table}}
}
\value{
character name of grouped class
}
\description{
Add necessary variables for robocop to layer table
}
\examples{
add_robocop_variables(get_layout_table())
}
\keyword{internal}
