% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_subtitle}
\alias{add_subtitle}
\title{Add subtitle to layout candidate}
\usage{
add_subtitle(
  robopptx,
  subtitle = NULL,
  layout = NA_integer_,
  shapeid = NA_integer_,
  hint = NULL
)
}
\arguments{
\item{robopptx}{robopptx Imported layout file from \link{load_layout}}

\item{sub_title}{character Subtitle to add to the current layout candidate}
}
\value{
A \code{robopptx} object.
}
\description{
Add subtitle to layout candidate
}
\examples{
"todo"
}
