% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_table}
\alias{add_table}
\title{Add Table to layout candidate}
\usage{
add_table(
  robopptx,
  table = NULL,
  layout = NA_integer_,
  shapeid = NA_integer_,
  hint = NULL
)
}
\arguments{
\item{robopptx}{robopptx Imported layout file from \link{load_layout}}

\item{table}{data.frame Table to add to the current layout candidate}
}
\value{
A \code{robopptx} object.
}
\description{
Add Table to layout candidate
}
\examples{
"todo"
}
