% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_text}
\alias{add_text}
\title{Add Text to layout candidate}
\usage{
add_text(
  robopptx,
  text = NULL,
  layoutid = NA_integer_,
  shapeid = NA_integer_,
  hint = NULL
)
}
\arguments{
\item{robopptx}{robopptx Imported layout file from \link{load_layout}}

\item{text}{character Text to add to the current layout candidate}
}
\value{
A \code{robopptx} object.
}
\description{
Add Text to layout candidate
}
\examples{
"todo"
}
