% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{can_robocop_handle}
\alias{can_robocop_handle}
\title{Verify that robocop can draw R-Object on PowerPoint}
\usage{
can_robocop_handle(x)
}
\arguments{
\item{x}{object}
}
\value{
character robocop content name
}
\description{
Verify that robocop can draw R-Object on PowerPoint
}
\examples{
can_robocop_handle("My Title")
}
\keyword{internal}
