% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{export}
\alias{export}
\title{Export populated slides to a PowerPoint presentation}
\usage{
export(robopptx, to = NULL)
}
\arguments{
\item{robopptx}{A \code{robopptx} object.}

\item{to}{TODO}
}
\value{
TODO
}
\description{
Export populated slides to a PowerPoint presentation
}
\examples{
"todo"
}
