% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{file_open}
\alias{file_open}
\title{Open local file in default application}
\usage{
file_open(path)
}
\arguments{
\item{path}{Path to file.}
}
\description{
Open local file in default application
}
\examples{
\dontrun{
file <- system.file("img", "Rlogo.png", package = "png")
file_open(file)
}

}
