% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robopptx.R
\name{get_emptyslide_table}
\alias{get_emptyslide_table}
\title{Get mapping table for PowerPoint types, robocop content and r classes}
\usage{
get_emptyslide_table(number_of_rows = 1)
}
\arguments{
\item{number_of_rows}{number of rows with default values}
}
\value{
data.table
}
\description{
Get mapping table for PowerPoint types, robocop content and r classes
}
\examples{
"todo"
}
\keyword{internal}
