% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robopptx.R
\name{get_layout_table}
\alias{get_layout_table}
\title{Get mapping table for PowerPoint types, robocop content and r classes}
\usage{
get_layout_table(rpptx, position_precision = 0.5)
}
\arguments{
\item{x}{object}
}
\value{
character name of grouped class
}
\description{
Get mapping table for PowerPoint types, robocop content and r classes
}
\examples{
"todo"
}
\keyword{internal}
