% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robopptx.R
\name{get_mapping_table}
\alias{get_mapping_table}
\title{Get mapping table for PowerPoint types, robocop content and r classes}
\usage{
get_mapping_table(supported_graph_classes = c("ggplot", "ms_chart"))
}
\arguments{
\item{supported_graph_classes}{object}
}
\value{
data.table data.table holding class mapping
}
\description{
Get mapping table for PowerPoint types, robocop content and r classes
}
\examples{
"todo"
}
\keyword{internal}
