% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{group_r_class}
\alias{group_r_class}
\title{Group common R classes for a PowerPoint placeholder mapping}
\usage{
group_r_class(x)
}
\arguments{
\item{x}{object}
}
\value{
character name of grouped class
}
\description{
Group common R classes for a PowerPoint placeholder mapping
}
\examples{
"todo"
}
\keyword{internal}
