% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{join_slides}
\alias{join_slides}
\title{Join all slide candidates into a presentation}
\usage{
join_slides(robopptx)
}
\arguments{
\item{robopptx}{robopptx Imported layout file from \link{load_layout}}
}
\value{
TODO
}
\description{
Join all slide candidates into a presentation
}
\examples{
"todo"
}
