% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{layout_duplicate_rename}
\alias{layout_duplicate_rename}
\title{Rename duplicated placeholders in PowerPoint Layout(s)}
\usage{
layout_duplicate_rename(x)
}
\arguments{
\item{x}{TODO}
}
\description{
Rename duplicated placeholders in PowerPoint Layout(s)
}
\examples{
"todo"
}
