% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{materialise}
\alias{materialise}
\title{Materialise current slide candidate into a slide}
\usage{
materialise(robopptx)
}
\arguments{
\item{robopptx}{robopptx Imported layout file from \link{load_layout}}
}
\value{
A \code{robopptx} object.
}
\description{
Materialise current slide candidate into a slide
}
\examples{
"todo"
}
