% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robopptx.R
\name{materialise_presentation}
\alias{materialise_presentation}
\title{Materialise all slides into a presentation}
\usage{
materialise_presentation(robopptx)
}
\arguments{
\item{robopptx}{robopptx Imported layout file from \link{load_layout}}
}
\value{
TODO
}
\description{
Materialise all slides into a presentation
}
\examples{
"todo"
}
