% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robopptx.R
\name{new_robopptx}
\alias{new_robopptx}
\title{Import and prepare PowerPoint file for robocop}
\usage{
new_robopptx(
  path = NULL,
  clean = "rename",
  delete_slides = FALSE,
  position_precision = 0.5
)
}
\value{
An object of class \code{"robopptx"}.
}
\description{
Load layout file, add slides with corresponding contents and export presentation.
}
\examples{
"todo"
obj <- new_robopptx()
}
