% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/officer_plot_layout_properties.R
\name{plot_layout_properties}
\alias{plot_layout_properties}
\title{Slide layout properties plot}
\usage{
plot_layout_properties(
  x,
  layout = NULL,
  master = NULL,
  labels = TRUE,
  title = TRUE,
  shape_id = TRUE
)
}
\arguments{
\item{x}{An \code{rpptx} object.}

\item{layout}{slide layout name to use}

\item{master}{master layout name where \code{layout} is located}

\item{labels}{if \code{TRUE}, placeholder labels will be printed, if \code{FALSE}
placeholder types and identifiers will be printed.}

\item{title}{if \code{TRUE}, a title with the layout name will be printed.}

\item{shape_id}{if \code{TRUE}, the shape id of the placeholder will be printed in the upper right
placeholder corner in blue.}
}
\description{
Plot slide layout properties and print informations into defined placeholders.
This can be useful to help visualise placeholders locations and identifier.
}
\examples{
x <- read_pptx()
plot_layout_properties(x = x, layout = "Title Slide", master = "Office Theme", title = TRUE, shape_id = TRUE)
}
\concept{functions for reading presentation informations}
\keyword{internal}
