% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{+.robopptx}
\alias{+.robopptx}
\title{Adding content with `+` operator}
\usage{
\method{+}{robopptx}(e1, e2)
}
\arguments{
\item{e1}{robopptx Imported layout file from \link{load_layout}}

\item{e2}{object Content to add with \link{add}}
}
\description{
Adding content with `+` operator
}
