% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{rclass_to_robo_content}
\alias{rclass_to_robo_content}
\title{Group common R classes for a PowerPoint placeholder mapping}
\usage{
rclass_to_robo_content(x)
}
\arguments{
\item{x}{object}
}
\value{
character robocop content name
}
\description{
Group common R classes for a PowerPoint placeholder mapping
}
\examples{
"todo"
}
\keyword{internal}
