% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{robocontent_dictionary}
\alias{robocontent_dictionary}
\title{Group common R classes for a PowerPoint placeholder mapping}
\usage{
robocontent_dictionary()
}
\arguments{
\item{x}{object}
}
\value{
character robocop content name
}
\description{
Group common R classes for a PowerPoint placeholder mapping
}
\examples{
robocontent_dictionary()
}
\keyword{internal}
