% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robocop-package.R
\docType{package}
\name{robocop-package}
\alias{robocop-package}
\title{robocop: Easily populate PowerPoint presentation using its layouts}
\description{
robocop is an externsion to the officer package to populate PowerPoint presentations very fast.
}
\author{
\strong{Maintainer}: Majid Eismann \email{yourself@somewhere.net}

Authors:
\itemize{
  \item Mark Heckmann \email{heckmann.mark@gmail.com} (\href{https://orcid.org/0000-0002-0736-7417}{ORCID})
}

}
\keyword{internal}
