% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robopptx.R
\name{slide_finalize}
\alias{slide_finalize}
\title{Internal slide function to select final layout}
\usage{
slide_finalize(slide, robocop_r6, selected_layout)
}
\arguments{
\item{robocop_r6}{robopptx Imported layout file from \link{load_layout}}

\item{selected_layout}{}
}
\value{
TODO
}
\description{
Internal slide function to select final layout
}
\examples{
"todo"
}
