% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robopptx.R
\name{slide_layout_candidate}
\alias{slide_layout_candidate}
\title{Internal slide function to select possible layout(s)}
\usage{
slide_layout_candidate(slide, robocop_r6)
}
\arguments{
\item{slide}{single slide to select layout for}

\item{robocop_r6}{R6}
}
\value{
TODO
}
\description{
Internal slide function to select possible layout(s)
}
\examples{
"todo"
}
