% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robopptx.R
\name{slide_match_placeholder}
\alias{slide_match_placeholder}
\title{Internal slide function to match users content and placeholders}
\usage{
slide_match_placeholder(slide, layout_candidate, robocop_r6)
}
\arguments{
\item{slide}{single slide to select layout for}

\item{robocop_r6}{R6}
}
\value{
TODO
}
\description{
Internal slide function to match users content and placeholders
}
\examples{
"todo"
}
