% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robopptx.R
\name{slide_pick_layout}
\alias{slide_pick_layout}
\title{Internal slide function to select final layout}
\usage{
slide_pick_layout(slide, matched_slide)
}
\arguments{
\item{slide}{single slide to select layout for}

\item{matched_slide}{}
}
\value{
TODO
}
\description{
Internal slide function to select final layout
}
\examples{
"todo"
}
