% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add}
\alias{add}
\title{Add content to layout candidate}
\usage{
add(
  robopptx,
  content = NULL,
  layoutid = NA_integer_,
  shapeid = NA_character_,
  hint = NULL,
  ...,
  robo_content = NA_character_,
  r_class = NA_character_
)
}
\arguments{
\item{robopptx}{robopptx Imported layout file from \link{load_layout}}

\item{content}{object Content to add to the current layout candidate}
}
\value{
A \code{robopptx} object.
}
\description{
Add content to layout candidate
}
\examples{
"todo"
}
