% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_footer}
\alias{add_footer}
\title{Add Text to layout candidate}
\usage{
add_footer(
  robopptx,
  footer = NULL,
  layout = NA_integer_,
  shapeid = NA_integer_,
  hint = NULL
)
}
\arguments{
\item{robopptx}{robopptx Imported layout file from \link{load_layout}}

\item{footer}{character Footer to add to the current layout candidate}
}
\value{
A \code{robopptx} object.
}
\description{
Add Text to layout candidate
}
\examples{
"todo"
}
