% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_graph}
\alias{add_graph}
\title{Add Graph to layout candidate}
\usage{
add_graph(
  robopptx,
  graph = NULL,
  layoutid = NA_integer_,
  shapeid = NA_integer_,
  hint = NULL,
  ...
)
}
\arguments{
\item{robopptx}{robopptx Imported layout file from \link{load_layout}}

\item{graph}{ggplot, mschart, charater Graph to add to the current layout candidate as ggplot, mschart or filepath to existing .img or .png file}
}
\value{
A \code{robopptx} object.
}
\description{
Add Graph to layout candidate
}
\examples{
"todo"
}
