% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_slide}
\alias{add_slide}
\title{Add content to slide candidate}
\usage{
add_slide(robopptx, ...)
}
\arguments{
\item{robopptx}{robopptx Imported layout file from \link{load_layout}}

\item{...}{object Content to add to the current slide candidate}
}
\value{
A \code{robopptx} object.
}
\description{
Add content to slide candidate
}
\examples{
"todo"
}
