% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_subtitle}
\alias{add_subtitle}
\title{Add subtitle to slide candidate}
\usage{
add_subtitle(robopptx, subtitle = NULL)
}
\arguments{
\item{robopptx}{robopptx Imported layout file from \link{load_layout}}

\item{sub_title}{character Subtitle to add to the current slide candidate}
}
\value{
A \code{robopptx} object.
}
\description{
Add subtitle to slide candidate
}
\examples{
"todo"
}
