% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_table}
\alias{add_table}
\title{Add Table to slide candidate}
\usage{
add_table(robopptx, table = NULL)
}
\arguments{
\item{robopptx}{robopptx Imported layout file from \link{load_layout}}

\item{table}{data.frame Table to add to the current slide candidate}
}
\value{
A \code{robopptx} object.
}
\description{
Add Table to slide candidate
}
\examples{
"todo"
}
