% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_text}
\alias{add_text}
\title{Add Text to slide candidate}
\usage{
add_text(robopptx, text = NULL)
}
\arguments{
\item{robopptx}{robopptx Imported layout file from \link{load_layout}}

\item{text}{character Text to add to the current slide candidate}
}
\value{
A \code{robopptx} object.
}
\description{
Add Text to slide candidate
}
\examples{
"todo"
}
