% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{add_title}
\alias{add_title}
\title{Add title to layout candidate}
\usage{
add_title(
  robopptx,
  title = NULL,
  layoutid = NA_integer_,
  shapeid = NA_integer_,
  hint = NULL
)
}
\arguments{
\item{robopptx}{robopptx Imported layout file from \link{load_layout}}

\item{title}{character Title to add to the current layout candidate}
}
\value{
A \code{robopptx} object.
}
\description{
Add title to layout candidate
}
\examples{
"todo"
}
