% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{join_slides}
\alias{join_slides}
\title{materialise / materialize current slide candidate into a slide}
\usage{
join_slides(robopptx)
}
\arguments{
\item{robopptx}{robopptx Imported layout file from \link{load_layout}}
}
\value{
TODO
}
\description{
materialise / materialize current slide candidate into a slide
}
\examples{
"todo"
}
