% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/layout.R
\name{layout_duplicate_delete}
\alias{layout_duplicate_delete}
\title{Delete duplicated placeholders in PowerPoint Layout(s)}
\usage{
layout_duplicate_delete(x, keep_max_id = TRUE)
}
\arguments{
\item{keep_max_id}{}
}
\description{
Delete duplicated placeholders in PowerPoint Layout(s)
}
