% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file.R
\name{load_layout}
\alias{load_layout}
\title{Load PowerPoint to extract its layout}
\usage{
load_layout(path = NULL, ...)
}
\arguments{
\item{...}{Passed on to \link{new_robopptx}.}

\item{filepath}{Path to PPTX file.}
}
\value{
A \code{robopptx} object.
}
\description{
Load PowerPoint to extract its layout
}
\examples{
"todo"
}
