% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{+.robopptx}
\alias{+.robopptx}
\title{Adding content with \code{+} operator}
\usage{
\method{+}{robopptx}(e1, e2)
}
\arguments{
\item{e1}{robopptx Imported layout file from \link{load_layout}}

\item{e2}{object Content to add with \link{add}}
}
\value{
A \code{robopptx} object.
}
\description{
Adding content with \code{+} operator
}
\keyword{internal}
