% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internals.R
\name{+,robopptx,ANY-method}
\alias{+,robopptx,ANY-method}
\title{Register an R6 class as an S4 class}
\usage{
\S4method{+}{robopptx,ANY}(e1, e2)
}
\arguments{
\item{e1}{An object of class 'robopptx'}

\item{e2}{object Content to add with \link{add}}
}
\value{
robopptx
}
\description{
This makes the R6 class "robopptx" available as an S4 class.
}
