% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/robopptx.R
\name{robocop}
\alias{robocop}
\title{R6 Class to manage instructions for populating PowerPoint presentation}
\description{
Load layout file, add slides with corresponding contents and export presentation.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-robocop-new}{\code{robocop$new()}}
\item \href{#method-robocop-show_layout}{\code{robocop$show_layout()}}
\item \href{#method-robocop-show_slide}{\code{robocop$show_slide()}}
\item \href{#method-robocop-show_slide_candidate}{\code{robocop$show_slide_candidate()}}
\item \href{#method-robocop-flush_candidate}{\code{robocop$flush_candidate()}}
\item \href{#method-robocop-join_slides}{\code{robocop$join_slides()}}
\item \href{#method-robocop-add_slide}{\code{robocop$add_slide()}}
\item \href{#method-robocop-explain}{\code{robocop$explain()}}
\item \href{#method-robocop-clone}{\code{robocop$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-robocop-new"></a>}}
\if{latex}{\out{\hypertarget{method-robocop-new}{}}}
\subsection{Method \code{new()}}{
Creates a new instance of this \link[R6:R6Class]{R6} class.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{robocop$new(rpptx, position_precision = 0.5)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{index}}{(\code{character(1)})\cr
Identifier of the object.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-robocop-show_layout"></a>}}
\if{latex}{\out{\hypertarget{method-robocop-show_layout}{}}}
\subsection{Method \code{show_layout()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{robocop$show_layout(layout)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-robocop-show_slide"></a>}}
\if{latex}{\out{\hypertarget{method-robocop-show_slide}{}}}
\subsection{Method \code{show_slide()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{robocop$show_slide(slide)}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-robocop-show_slide_candidate"></a>}}
\if{latex}{\out{\hypertarget{method-robocop-show_slide_candidate}{}}}
\subsection{Method \code{show_slide_candidate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{robocop$show_slide_candidate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-robocop-flush_candidate"></a>}}
\if{latex}{\out{\hypertarget{method-robocop-flush_candidate}{}}}
\subsection{Method \code{flush_candidate()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{robocop$flush_candidate()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-robocop-join_slides"></a>}}
\if{latex}{\out{\hypertarget{method-robocop-join_slides}{}}}
\subsection{Method \code{join_slides()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{robocop$join_slides()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-robocop-add_slide"></a>}}
\if{latex}{\out{\hypertarget{method-robocop-add_slide}{}}}
\subsection{Method \code{add_slide()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{robocop$add_slide(content_list = list())}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-robocop-explain"></a>}}
\if{latex}{\out{\hypertarget{method-robocop-explain}{}}}
\subsection{Method \code{explain()}}{
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{robocop$explain()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-robocop-clone"></a>}}
\if{latex}{\out{\hypertarget{method-robocop-clone}{}}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{robocop$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
