% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add.R
\name{slide,robopptx-method}
\alias{slide,robopptx-method}
\title{Method for the 'myFunction' generic for objects of class 'MyClass'}
\usage{
\S4method{slide}{robopptx}(x)
}
\arguments{
\item{robopptx}{robopptx Imported layout file from \link{load_layout}}
}
\value{
A \code{robopptx} object.
}
\description{
Method for the 'myFunction' generic for objects of class 'MyClass'
}
